Update README.md
    Problem 1: Getting Started with the LC-22

In this homework, you will be using the LC-22 ISA to complete a Tower of Hanoi move-counting function. Before you begin, you should familiarize yourself with the available instructions, the register conventions and the calling convention of LC-22. Details can be found in the section, Appendix A: LC-22 Instruction Set Architecture, at the end of this document.

The assembly folder contains several tools for you to use:

    assembler.py: a basic assembler that can take your assembly code and convert it into binary instruc-tions for the LC-22.

    lc22.py: the ISA definition file for the assembler, which tells assembler.py the instructions supported by the LC-22 and their formats.

    lc22-sim.py: A simulator of the LC-22 machine. The simulator reads binary instructions and emulates the LC-22 machine, letting you single-step through instructions and check their results.